front/carrousel: module items a key prop added
